+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds the `"actionIdentity"`, `"actionIdentityFilled"`, `"actionSearch"`, `"mapsPlace"`, `"socialSpotify"`, and `"socialVimeo"` options to the `Icon` component.
+
 ### Updates
 
 - Updates `List` styling so the bottom border on description lists matches other borders.